backend/pyproject.toml
[project]
name = "tahrix-backend"
version = "0.1.0"
description = "TAHRIX — Agentic AI Blockchain Cyber Intelligence (AABCI) Backend"
authors = [{ name = "Fikri Armia Fahmi", email = "fikri@example.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Config & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    # HTTP client
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    # DB: PostgreSQL
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # DB: Neo4j
    "neo4j>=5.26.0",
    # Cache + queue
    "redis>=5.2.0",
    "celery>=5.4.0",
    # Auth & security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.9.0",
    # Blockchain helpers
    "eth-utils>=5.1.0",
    "eth-hash[pycryptodome]>=0.7.0",
    "base58>=2.1.1",
    # ML / inference (kept minimal in API; training in ml/)
    "onnxruntime>=1.20.0",
    "numpy>=1.26.0,<2.2.0",
    # Reporting
    "weasyprint>=63.0",
    "jinja2>=3.1.4",
    "python-docx>=1.1.0",
    "markdown>=3.7",
    "lxml>=5.3.0",
    # IPFS
    "requests-toolbelt>=1.0.0",
    # LLM (provider-agnostic; we adapt OpenAI-compatible)
    "openai>=1.54.0",
    # Observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "respx>=0.21.1",
    "locust>=2.32.0",
]
ml = [
    "torch>=2.6.0",
    "torch-geometric>=2.5.0",
    "shap>=0.46.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "onnx>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "ASYNC"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
